Add to MRI_FILE_PREPROCESS script a check to not override existing files, and support flair scans

+++ tim_analysis/mri_file_preprocess.py
@@ -88,36 +88,58 @@ if os.path.exists("./FIELDMAP"):
         os.mkdir(fmap_path)
     # NOTE - we're combining two magnitudes into 1. If we don't want to do so, we need to remove '-m y' from the command
     # and catch this case in the naming as well (as magnitude2).
+    current_files_in_path = os.listdir(fmap_path)
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -m y -p y -z y -o {fmap_path} ./FIELDMAP', shell=True))
     for file in os.listdir(fmap_path):
-        print(f"Handling file {file}")
-        suffix = get_suffix(file)
-        new_name = f"{subject}_{session}_run-{1 if '_PA' in file else 2}"
-        if "_ph." in file:
-            new_name += "_phasediff"
-        else:
-            new_name += "_magnitude1"
-        new_name += f".{suffix}"
-        print(f"Renaming file to {new_name}")
-        os.rename(f"{fmap_path}/{file}", f"{fmap_path}/{new_name}")
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            new_name = f"{subject}_{session}_run-{1 if '_PA' in file else 2}"
+            if "_ph." in file:
+                new_name += "_phasediff"
+            else:
+                new_name += "_magnitude1"
+            new_name += f".{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{fmap_path}/{file}", f"{fmap_path}/{new_name}")
 else:
     print("No Fieldmap files to process. Moving on.")
 
 # Handle T1 files
 if os.path.exists("./T1"):
     print("Starting conversion of T1 scans")
-    t1_path = f"./{session}/anat"
-    if not os.path.exists(t1_path):
-        os.mkdir(t1_path)
-    print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {t1_path} ./T1', shell=True))
-    for file in os.listdir(t1_path):
-        print(f"Handling file {file}")
-        suffix = get_suffix(file)
-        new_name = f"{subject}_{session}_T1w.{suffix}"
-        print(f"Renaming file to {new_name}")
-        os.rename(f"{t1_path}/{file}", f"{t1_path}/{new_name}")
+    anat_path = f"./{session}/anat"
+    if not os.path.exists(anat_path):
+        os.mkdir(anat_path)
+    current_files_in_path = os.listdir(anat_path)
+    print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {anat_path} ./T1', shell=True))
+    for file in os.listdir(anat_path):
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            new_name = f"{subject}_{session}_T1w.{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{anat_path}/{file}", f"{anat_path}/{new_name}")
 else:
     print("No T1 files to process. Moving on.")
+
+# Handle ANATOMY files
+if os.path.exists("./ANATOMY"):
+    print("Starting conversion of FLAIR scans")
+    anat_path = f"./{session}/anat"
+    if not os.path.exists(anat_path):
+        os.mkdir(anat_path)
+    current_files_in_path = os.listdir(anat_path)
+    print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {anat_path} ./T1', shell=True))
+    for file in os.listdir(anat_path):
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            new_name = f"{subject}_{session}_FLAIR.{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{anat_path}/{file}", f"{anat_path}/{new_name}")
+else:
+    print("No FLAIR files to process. Moving on.")
 
 #Handle DTI files
 if os.path.exists("./DTI"):
@@ -125,6 +147,7 @@ if os.path.exists("./DTI"):
     dwi_path = f"./{session}/dwi"
     if not os.path.exists(dwi_path):
         os.mkdir(dwi_path)
+    current_files_in_path = os.listdir(dwi_path)
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {dwi_path} ./DTI', shell=True))
     suffix_numbers = defaultdict(int)
     # A DTI run should produce 4 files - nifti, json, bval and bvec.
@@ -132,16 +155,17 @@ if os.path.exists("./DTI"):
     for file in os.listdir(dwi_path):
         suffix_numbers[file.split(".")[0].split("_")[-1]] += 1
     for file in os.listdir(dwi_path):
-        suffix = get_suffix(file)
-        old_name = file.split(".")[0]
-        if suffix_numbers[old_name.split("_")[-1]] == 4:
-            print(f"handling file {file}")
-            new_name = f"{subject}_{session}_dwi_{'pa' if '_PA_' in old_name else 'ap'}.{suffix}"
-            print(f"renaming file to {new_name}")
-            os.rename(f"{dwi_path}/{file}", f"{dwi_path}/{new_name}")
-        else:
-            print(f"file {file} doesn't have 4 of the same enumaration ({old_name.split('_')[-1]}). Deleting.")
-            os.remove(f"{dwi_path}/{file}")
+        if file not in current_files_in_path:
+            suffix = get_suffix(file)
+            old_name = file.split(".")[0]
+            if suffix_numbers[old_name.split("_")[-1]] == 4:
+                print(f"handling file {file}")
+                new_name = f"{subject}_{session}_dwi_{'pa' if '_PA_' in old_name else 'ap'}.{suffix}"
+                print(f"renaming file to {new_name}")
+                os.rename(f"{dwi_path}/{file}", f"{dwi_path}/{new_name}")
+            else:
+                print(f"file {file} doesn't have 4 of the same enumaration ({old_name.split('_')[-1]}). Deleting.")
+                os.remove(f"{dwi_path}/{file}")
 else:
     print("No DTI files to process. Moving on.")
 
@@ -151,22 +175,24 @@ if os.path.exists("./REST"):
     func_path = f"./{session}/func"
     if not os.path.exists(func_path):
         os.mkdir(func_path)
+    current_files_in_path = os.listdir(func_path)
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {func_path} ./REST', shell=True))
     for file in os.listdir(func_path):
-        print(f"Handling file {file}")
-        suffix = get_suffix(file)
-        echo = get_echo(file)
-        new_name = f"{subject}_{session}_task-rest_{echo}_bold.{suffix}"
-        print(f"Renaming file to {new_name}")
-        os.rename(f"{func_path}/{file}", f"{func_path}/{new_name}")
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            echo = get_echo(file)
+            new_name = f"{subject}_{session}_task-rest_{echo}_bold.{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{func_path}/{file}", f"{func_path}/{new_name}")
 else:
     print("No RS files to process. Moving on.")
 
 # Handle TIM Files
-# MAKE SURE FOLDER IS IN FORMAT WARX AND NOT WAR_X
+# MAKE SURE FOLDER IS IN FORMAT TIMX AND NOT TIM_X
 for tim_run in range(1, runs + 1):
-    malformatted_tim_folder = f"./WAR {tim_run}"
-    tim_folder = f"./WAR{tim_run}"
+    malformatted_tim_folder = f"./TIM {tim_run}"
+    tim_folder = f"./TIM{tim_run}"
     if os.path.exists(malformatted_tim_folder):
         os.rename(malformatted_tim_folder, tim_folder)
     if os.path.exists(tim_folder):
@@ -201,7 +227,7 @@ for current_file in os.listdir():
         else:
             print(f"Could not find block number in {current_file}. Skipping.")
             continue
-        df = pd.read_csv(current_file)
+        df = pd.read_csv(current_file, delimiter='\t')
         df.drop("Unnamed: 0", axis=1, inplace=True)
         df = df.round({"Time": 2, "Duration": 2})
         new_file_name = f"{subject}_{session}_task-tim_run-{run_number}_events.tsv"
