+++ war_analysis/mri_file_preprocess.py
@@ -77,20 +77,22 @@ if os.path.exists("./FIELDMAP"):
     fmap_path = f"./{session}/fmap"
     if not os.path.exists(fmap_path):
         os.mkdir(fmap_path)
+    current_files_in_path = os.listdir(fmap_path)
     # NOTE - we're combining two magnitudes into 1. If we don't want to do so, we need to remove '-m y' from the command
     # and catch this case in the naming as well (as magnitude2).
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -m y -p y -z y -o {fmap_path} ./FIELDMAP', shell=True))
     for file in os.listdir(fmap_path):
-        print(f"Handling file {file}")
-        suffix = get_suffix(file)
-        new_name = f"{subject}_{session}_run-{1 if '_PA' in file else 2}"
-        if "_ph." in file:
-            new_name += "_phasediff"
-        else:
-            new_name += "_magnitude1"
-        new_name += f".{suffix}"
-        print(f"Renaming file to {new_name}")
-        os.rename(f"{fmap_path}/{file}", f"{fmap_path}/{new_name}")
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            new_name = f"{subject}_{session}_run-{1 if '_PA' in file else 2}"
+            if "_ph." in file:
+                new_name += "_phasediff"
+            else:
+                new_name += "_magnitude1"
+            new_name += f".{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{fmap_path}/{file}", f"{fmap_path}/{new_name}")
 else:
     print("No Fieldmap files to process. Moving on.")
 
@@ -100,15 +102,35 @@ if os.path.exists("./T1"):
     t1_path = f"./{session}/anat"
     if not os.path.exists(t1_path):
         os.mkdir(t1_path)
+    current_files_in_path = os.listdir(t1_path)
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {t1_path} ./T1', shell=True))
     for file in os.listdir(t1_path):
-        print(f"Handling file {file}")
-        suffix = get_suffix(file)
-        new_name = f"{subject}_{session}_T1w.{suffix}"
-        print(f"Renaming file to {new_name}")
-        os.rename(f"{t1_path}/{file}", f"{t1_path}/{new_name}")
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            new_name = f"{subject}_{session}_T1w.{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{t1_path}/{file}", f"{t1_path}/{new_name}")
 else:
     print("No T1 files to process. Moving on.")
+
+# Handle ANATOMY files
+if os.path.exists("./ANATOMY"):
+    print("Starting conversion of FLAIR scans")
+    anat_path = f"./{session}/anat"
+    if not os.path.exists(anat_path):
+        os.mkdir(anat_path)
+    current_files_in_path = os.listdir(anat_path)
+    print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {anat_path} ./T1', shell=True))
+    for file in os.listdir(anat_path):
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            new_name = f"{subject}_{session}_FLAIR.{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{anat_path}/{file}", f"{anat_path}/{new_name}")
+else:
+    print("No FLAIR files to process. Moving on.")
 
 #Handle DTI files
 if os.path.exists("./DTI"):
@@ -116,6 +138,7 @@ if os.path.exists("./DTI"):
     dwi_path = f"./{session}/dwi"
     if not os.path.exists(dwi_path):
         os.mkdir(dwi_path)
+    current_files_in_path = os.listdir(dwi_path)
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {dwi_path} ./DTI', shell=True))
     suffix_numbers = defaultdict(int)
     # A DTI run should produce 4 files - nifti, json, bval and bvec.
@@ -123,16 +146,17 @@ if os.path.exists("./DTI"):
     for file in os.listdir(dwi_path):
         suffix_numbers[file.split(".")[0].split("_")[-1]] += 1
     for file in os.listdir(dwi_path):
-        suffix = get_suffix(file)
-        old_name = file.split(".")[0]
-        if suffix_numbers[old_name.split("_")[-1]] == 4:
-            print(f"handling file {file}")
-            new_name = f"{subject}_{session}_dwi_{'pa' if '_PA_' in old_name else 'ap'}.{suffix}"
-            print(f"renaming file to {new_name}")
-            os.rename(f"{dwi_path}/{file}", f"{dwi_path}/{new_name}")
-        else:
-            print(f"file {file} doesn't have 4 of the same enumaration ({old_name.split('_')[-1]}). Deleting.")
-            os.remove(f"{dwi_path}/{file}")
+        if file not in current_files_in_path:
+            suffix = get_suffix(file)
+            old_name = file.split(".")[0]
+            if suffix_numbers[old_name.split("_")[-1]] == 4:
+                print(f"handling file {file}")
+                new_name = f"{subject}_{session}_dwi_{'pa' if '_PA_' in old_name else 'ap'}.{suffix}"
+                print(f"renaming file to {new_name}")
+                os.rename(f"{dwi_path}/{file}", f"{dwi_path}/{new_name}")
+            else:
+                print(f"file {file} doesn't have 4 of the same enumaration ({old_name.split('_')[-1]}). Deleting.")
+                os.remove(f"{dwi_path}/{file}")
 else:
     print("No DTI files to process. Moving on.")
 
@@ -142,14 +166,16 @@ if os.path.exists("./REST"):
     func_path = f"./{session}/func"
     if not os.path.exists(func_path):
         os.mkdir(func_path)
+    current_files_in_path = os.listdir(func_path)
     print(subprocess.run(f'{DCM_CONVERTER_PATH} -f "%p_%s" -p y -z y -o {func_path} ./REST', shell=True))
     for file in os.listdir(func_path):
-        print(f"Handling file {file}")
-        suffix = get_suffix(file)
-        echo = get_echo(file)
-        new_name = f"{subject}_{session}_task-rest_{echo}_bold.{suffix}"
-        print(f"Renaming file to {new_name}")
-        os.rename(f"{func_path}/{file}", f"{func_path}/{new_name}")
+        if file not in current_files_in_path:
+            print(f"Handling file {file}")
+            suffix = get_suffix(file)
+            echo = get_echo(file)
+            new_name = f"{subject}_{session}_task-rest_{echo}_bold.{suffix}"
+            print(f"Renaming file to {new_name}")
+            os.rename(f"{func_path}/{file}", f"{func_path}/{new_name}")
 else:
     print("No RS files to process. Moving on.")
